Update: Add user and user Routes

+++ src/controller/todo.ts
@@ -1,35 +1,42 @@
 import { Request, Response } from "express";
-import * as TodoService from "../services//todo";
+import * as TodoService from "../services/todo";
+import { UUID } from "crypto";
 
-export function getTodos(_: Request, res: Response) {
-  const data = TodoService.getTodos();
+export async function getTodos(_: Request, res: Response) {
+  const data = await TodoService.getTodos();
   if (!data) {
     res.status(404).json({
       message: "No todos found",
     });
   } else {
-    res.status(404).json([...data]);
+    res.status(200).json([...data]);
   }
 }
 
 export function createTodo(req: Request, res: Response) {
   const { body } = req;
   if (!req.body.task) {
-    res.status(404).send({
+    res.status(404).json({
       message: "Task is missing. Couldn't create a Todo",
     });
   }
   const { task } = body;
   const service_response = TodoService.createTodo(task as string);
   if (service_response.data) {
-    res.status(200).send(service_response);
+    res.status(200).json(service_response);
   } else {
-    res.status(404).send(service_response);
+    res.status(404).json(service_response);
   }
 }
 
-export function updateTodo(req: Request, res: Response) {
-  const id = parseInt(req.params.id, 10);
+export function updateTodo(req: Request<{ id?: UUID }>, res: Response) {
+  const { id } = req.params;
+  if (!id) {
+    res.status(404).json({
+      error: "Id is required",
+    });
+    return;
+  }
   const { query } = req;
   const { task, isCompleted } = req.body;
   const service_response = TodoService.updateTodo(
@@ -39,20 +46,24 @@ export function updateTodo(req: Request, res: Response) {
     isCompleted,
   );
   if (service_response.data) {
-    res.status(200).send(service_response);
+    res.status(200).json(service_response);
   } else {
-    res.status(404).send(service_response);
+    res.status(404).json(service_response);
   }
 }
 
-export function deleteTodo(req: Request, res: Response) {
-  const { params } = req;
-  const id = parseInt(params.id, 10);
-  console.log(id);
+export function deleteTodo(req: Request<{ id?: UUID }>, res: Response) {
+  const { id } = req.params;
+  if (!id) {
+    res.status(404).json({
+      error: "Id is required",
+    });
+    return;
+  }
   const service_response = TodoService.deleteTodo(id);
   if (service_response.data) {
-    res.status(200).send(service_response);
+    res.status(200).json(service_response);
   } else {
-    res.status(404).send(service_response);
+    res.status(404).json(service_response);
   }
 }
